Solve warings in const.

+++ src/const.cpp
@@ -1,17 +1,9 @@
 #include "const.h"
 
-int oneintlen = sizeof(int);
-int twointlen = 2*sizeof(int);
-int threeintlen = 3*sizeof(int);
+int oneintlen = (int)sizeof(int);
+int twointlen = 2*(int)sizeof(int);
+int threeintlen = 3*(int)sizeof(int);
 
-int oneptrlen = sizeof(char*);
+int oneptrlen = (int)sizeof(char*);
 
-//int kalign=ALIGNK;
-//int valign=ALIGNV;
-//int talign=MAX(kalign, valign);
-
-
-//int kalignm=kalign-1;
-//int valignm=valign-1;
-//int talignm=talign-1;
